Move const to gsm class
