--- appium/webdriver/gsm_signal_strength.py
@@ -1,21 +1,0 @@
-#!/usr/bin/env python
-
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-
-class GsmSignalStrength(object):
-    NONE_OR_UNKNOWN = 0
-    POOR = 1
-    MODERATE = 2
-    GOOD = 3
-    GREAT = 4
